Added EFL and Elementary Core Libraries and Terminology
These formulae have been tested to work on OS X 10.9.4 with XQuartz 2.7.8_beta1
(xorg-server 1.16.0).

EFL, Elementary and Terminology are all compiled based on the document found
on the Enlightenment wiki, https://phab.enlightenment.org/w/osx/.

Terminology runs as expected, however there is no sound. That may be due to a
configuration issue with PulseAudio, but I'm not sure.